Extract ResetContextStorage into top-level test helper.
